[symbolize] Cache llvm-symbolizer results
This change handles the spew case that causes critical slow
downs when processing the symbolizer spew from bad asan output.
Right now on 16MB inputs mostly filled with backtraces the filter
is grinding to a halt as a result of all the backtracing. This
change should massively speed up that case.

INTK-441

Change-Id: I90ec053fdbe6c66252486a7981e5a9e3237a76ba