[ledger] Append 'local' to test-suite names
This depends on https://fuchsia-review.googlesource.com/c/garnet/+/177419
which consumes the test suite name. It's still safe to submit this first.

This will not yet change the name of the test suite that appears in
the dashboard, because //peridot/tests/benchmarks/peridot_benchmarks.sh
overrides the test suite name via a command line argument.  In the future,
we should switch to reading the test suite name and the test name from the
tspec file, and toss the command line flag, since the tspec file
is already a natural place to communicate this information.

This change is in preperation for that.

Test: On CQ

IN-422

Change-Id: Idc353f3bf50ba37e62e5d2132cd3bb1d3febffa7