[amber] Add logging library
Amber needs to be able to log to the fuchsia logger,
and a logging file, while also allowing the libraries
to be shared between the host and target. This library
wraps all the possible logging targets with an interface
to make sure it's portable.

CP-106 ought to address this in a more robust way inside
fuchsia, but this still might be useful when we need to
share logging across the host and target.

TEST=updated unit tests, added lots of manual testing

Change-Id: Ic21775938bc0b248f0f58517f9a6cdbb08d2d8f2