Revert "[products] introduce "products" to describe builds"
This change reverts https://fuchsia-review.googlesource.com/c/build/+/176743

Reason: build artifacts were altered, which somehow resulted in the failure
of a CI-only step to tar up amber-files/repository ('tar fuchsia packages').

Tested: the above issue was bisected to the third-party roll of this
(and one other change) into Garnet. A swarming task was spun up to simulate
running this revert in CI - and it passed:
https://ci.chromium.org/swarming/task/3ee9429e76951710?server=chromium-swarm.appspot.com

Change-Id: I13627d34154a5becb9950b072cfd1a92911edbe1