[tee][optee] Add open session interface
To communicate between a client app and trusted app pair, a session must
be created. This is done by sending an OP-TEE message to the secure
world requesting a new session be established. This CL adds an interface
to the OP-TEE device driver for opening a session and receiving a
session id back. That id is then stored in the implementation specific
portion of the TEE Client API's TEEC_Session.

In practice, this doesn't quite work yet. Upon receiving the session
request, OP-TEE replies with an RPC request for shared memory. Since we
haven't added RPC handling yet, we fail there.

Test: Run tee-test and see that it fails when RPC request is made by
secure world.

Change-Id: I56ff5f2a98b73c3d262738197b6bb62e2ef58fa5