[tee] Rename SMC Message to SMC Function
Per ARM SMCCC nomenclature, a SMC call targets a function of a secure
monitor provided service. As such, the SMC parameters that make up an
SMC call should be referred to as a function call rather than a message.
This will also help disambiguate from OP-TEE messages, which are an
optional parameters in shared memory.

This change is just a rename, no functionality changed.

Test: Compilation successful
Change-Id: Ieca80e27c777d07ed42997ba643201424ec0dbc2