[utest][perftest] Don't copy compiler-generated code to another process
Copying test code to another process has precise requirements.  It's
never safe to assume that compiler-generated code can be sliced out
of one process and pasted into another one that doesn't share the
rest of the original process's code.  It's a fragile approach that
should always be avoided.  When it's necessary to inject code, it
should be separately compiled or hand-composed in assembly.

In perftest, there was never any real need to copy code into the
test process.  The vDSO entry point itself can be the entry point
for the process.  The fragile method in this case in the ASan built
resulted in copying some instrumented code that could never work in
a process that isn't fully set up with the ASan runtime in place.

Test: CQ + manual test with USE_ASAN=true build with new compiler
Change-Id: I088e1f499f243317503ae25adb84f20725b11088