[netboot][zbi] Support complete ZBI booting
When netboot receives only a "kernel.bin" and no "ramdisk.bin",
it expects the single file to be a complete ZBI.  It splits the
kernel from the data ZBI and passes both to mexec.

The BOOTLOADER_VERSION is bumped to 0.7.9 for this change, to
indicate a netboot that's capable of booting a complete ZBI.
This means that with this change, new and old zedboot images and
host-side bootserver tools will be incompatible.  (That is,
bootserver will refuse to serve a differently-versioned zedboot
even though in this case it would work as long as either the old
method with separate ramdisk image was being used, or the zedboot
was a new one.)

Test: CQ + manual netboot with both old and new methods
Change-Id: I63ff688e491e5f956ffc46f7284f613bad2d3633