[host][zbi] Use std::byte
In C++ 17 std::byte is the most kosher official thing to use.
uint8_t happens to be unsigned char in practice, which is also
kosher in C++ but std::byte is preferred and there's no standard
guarantee that uint8_t is actually unsigned char and thus has the
special status for C++ aliasing rules.

Test: CQ
Change-Id: I3bf69689155a451ab4df38ea1bb9fc6cdd2a02e5