[mediaplayer] replace fuchsia.mediaplayer.MediaMetadata definition
This CL replaces fuchsia.mediaplayer.MediaMetadata with
fuchsia.mediaplayer.Metadata. MediaMetadata was a structure with a
fixed set of string-typed fields for the various properties and
a numeric 'duration' field.

Metadata contains a vector of Property structs, which each consist
of a string label and string value. This makes metadata extensible.

Duration is no longer part of the metadata representation, and is
represented separately in the MediaPlayerStatus struct.

The internal representation of metadata is now just an unordered_map
of strings to strings, which has the alias media_player::Metadata.

TEST: No behavior change.

Change-Id: Iad2e1e464b0d596780b89a8888d02c79df4507aa