[runtests] Forward the environment variable "RUNTESTS_TEST_CLASS" on POSIX.
This environment variable is set in run-all-tests.cpp, so not forwarding
it means runtests didn't have the same behavior on POSIX as on Fuchsia,
which is not intended.

Test: Ran trace-reader-test through runtests on my linux machine.
Before this change, the test size flags had no effect.
After this change, running with "-s" results in all the test
cases being skipped (because they're medium).

Change-Id: I35585ebdd360dd7a5fecb685ccb225b85d952edc