[rust] Don't wake up for defunct timers
Often, one has a future that completes when either
    A) Some event happens (e.g., a message is received);
    B) A timeout occurs.

If A happens first, then the timeout future is dropped but its waker is
not cleaned up from the heap in the executor, which means that we will
wake up for that timeout no matter what. This change addresses the issue
by storing timer wakers by weak reference and throwing out dropped wakers
from the heap when computing the next deadline for zx_port_wait.

The primary motivation for this change is ease of testing using
wake_next_timer(): we'd like it to not return timeouts for dropped
timers.

TEST=Existing tests in timer.rs should be sufficient

Change-Id: I1f849f578099bed8e64b36defe947ee3e50d8e3c