[products] Introduce "products" to describe builds
A "product" definition is stored under {layer}/packages/products. A product
definition is a JSON file containing and object with the following fields:

monolith:
  a list of packages included in OTA images, base system images, and the
  distribution repository.
preinstall:
  a list of packages pre-installed on the system (also added to the
  distribution repository)
available:
  a list of packages only added to the distribution repository)

Note: preinstall does not fully work as yet. The blobs are added to the
blobfs, however, no dynamic index is added to the minfs image that would
indicate to pkgsvr that they are preinstalled. This feature requires changes
in Garnet. Having the blobs pre-seeded in blobfs is still useful as it may
greatly speed up time to activation of those packages.

Additional packages specified via fuchsia_packages or synthesize_packages are
at this time added to monolith_packages, but should be demoted to
preinstall_packages once it is fully working.

Test: manual builds of garnet, peridot, topaz. CQ, and manual CI test.

This reverts commit fc28766546bcc015a74cef15e9ade453758d9638.

Change-Id: I6e0c26cfbf04a33a546a09b929775e43c012f6df