[Codec][amlogic-video] use_h264_decoder_test pass
Initial connection from Codec interface to AmlogicVideo, so that the new
manual use_h264_decoder_test passes when run on VIM2, with some caveats.

The layers are CodecImpl to handle protocol enforcement and de-async-ing
stream switch handling, then the in-process mostly-sync CodecAdapter
interface implemented by codec_adapter_h264.cc, which in turn talks to
AmlogicVideo and related to get some bear.h264 frames decoded via the HW
decoder.

CodecImpl is a separate layer from CodecAdapter because CodecImpl will
be usable by other Codec servers while CodecAdapter implementations
typically will be unique to a given Codec server. The CodecImpl and
directly-related code is based partly on omx_codec_runner code, but
refactored to be usable for amlogic_video_decoder driver, and mostly
refactored to be usable for omx_codec_runner also.  For now this code is
in amlogic_video_decoder's directory since it's not quite ready to be a
separate source_set yet.

Up to this point we've been avoiding creating merge conflict in
AmlogicVideo.  With this CL merged, we can make some more-significant
but tighter-scoped (than this CL) changes to AmlogicVideo to allow
things to line up better between codec_adapter_h264.cc and
amlogic-video.cc, to be more efficient, and to remove the various other
caveats.

This CL doesn't imply that it's worthwhile to try general h264 decoding;
we'll need to fix a few of the larger caveats before that'll work well.

Tested: use_h264_decoder_test manually run on VIM2
Change-Id: I83361af77770862c9246ce8afa983b8350f7fc2b