[fx][boot] Add ability to flash bootloader
Gauss and astro both have the ability to pave bootloader images. A flag
was added to expose this capability to fx boot for conveniance. It is
necessary for the image to be prebuilt since the bootloader source is not
part of any Fuchsia repo.

Change-Id: I6933aebab84764d84a094859af0f813b4badfa2e
Tested: Paved a bootloader to astro.