[zircon] Forward configs, but not deps, from Zircon shared libraries
When generating a build target for a shared library from Zircon we want
to propagate the configuration information from dependencies of that
library, but not the deps. This way targets can use header files from
a shared library they depend on, including relative #includes they may
have, but do not have to link against things the shared library depends
on.

BLD-224 #comment
Test: CQ

Change-Id: I38ca564294024445afba7fbed7de377f01a6a6f3