[debugger] Add a symbol context for relative addrs.
Passing this encodes the load address and allows the ModuleSymbols class
to deal in absolute addresses with its API.

This eliminates the need for the complicated LoadedModuleSymbols
pass-through API. That class is now a simple holder for a
ModuleSymbols reference and the new SymbolContext helper class.

In the future, the symbol objects like Function and CodeBlock will
require a SymbolContext parameter any time they deal with addresses so
the interface can be in absolute addresses, while the DWARF symbol
objects can continue to be independent of load address.

TEST=unit

Change-Id: I10c82a382f4e25315bd2780f2b48a247c450c110