[bt][rfcomm] Allow certain pre-mux-startup frames
Implement the limited set of valid pre-multiplexer-startup frames.
The original design of Frame assumed that every Frame would have a valid
Role specified. This is not the case; the Frames needed to start the
multiplexer will not have a valid (i.e. kInitiator or kResponder) Role
attached, as the multiplexer startup procedure is what determines RFCOMM
roles.

NET-1122 #comment
TEST=bt-host-unittests
--gtest_filter=WritePreMuxStartup*,ReadFramePreMuxStartup

Change-Id: I518f71c5a92e1b2d1f06131f2dc65df3366033a8