[audio] Clarify delay and token
This CL addresses two TODOs that resulted from the previous CL.
Specifically, the processing algorithm for Delay has been
clarified (and optimized), and the dfx_token concept has been
made into an actual typedef that represents a void* (rather
than a uint64_t). An additional test of the DfxDelay ensures
that it operates correctly at the outer limits allowed.

Test: includes audio_dfx_tests updates; all still pass.

Change-Id: Iabee7402c63912f917dca2ea58b93129bbd196b6