[kernel][exception] Add zx_status_t string printing to exception assert
To add zx_status_t to string conversion to the exception assert and
potentially other places in the kernel we move the current function in
ulib to the kernel.  Since ulib still needs to export the convertion,
we don't want to duplicate the definition in 2 places, and we don't
want to add a new syscall for this, we add the conversion as a static
inline function in an 'impl' header file and use it in source files in
both the kernel and ulib.

ZX-2373 #done Add zx_status string printing to exception assert

Test: out of memory exception shows expected assertion status string

Change-Id: I4ea4226ea848a2a188a79317efe5f5c1f8f971c4