[appmgr] Determine runtime before reading binary
Currently we check for bin/app if we cannot read runtime metadata.
Consolidate to one place where we determine the runner, and then branch
off to either using the runner, or falling back to ELF binaries (can be
thought as the "default runner").

CP-19 #comment

TESTED=ran all layers, observed it works.
/pkgfs/packages/appmgr_unittests/0/test/appmgr_unittests
  passes

Change-Id: Icaf009d36cc91b453774b0605314d26e07080810