[garnet] wlan-smoke-test: add command line options
Add command line args for the target ssid, target password
and a flag indicating if the connection should be forced
if the interface is already connected to the target network.
For the optional password:"" is used as the default (and
for open networks).

usage:
wlan-smoke-test <target_ssid> [-p <target_pwd>] [-s]

Bug: NET-1095
Test: build and boot garnet, manually confirm output
Test: fx run-test wlan_smoke_test_bin_test_rustc
Change-Id: I630c32fd6594436fc62334bab0a2e8b6e9579043