[set] reintroduce the layer guess to fix --netboot
I had removed this during an iteration of product integration of //build
where it worked removed, however, in order to ease the transition through
infrastructure I made default product selection in GN contingent on an empty
package list as well as an empty product list. The side effect is that we
must retain this special workaround for the --netboot convenience.

Change-Id: I9e3c1aecc88b2ad72fe41d1251851170985a2554