[uapp] gfxlatency version 2
This makes some significant changes to this app in order to have
it be more useful for latency measurements.

The app simulates a desktop shell with 2 clients:

1. Drawing canvas that can be scrolled.
2. Cursor sprite.

The --slow-down-scale-factor flag can be used to simulate
a slower drawing canvas client. The app can be used to
determine what the minimal latency is for a gfx client and
demonstrates how you can potentially have a slow compositor
while still providing very low latency updates for some
clients under the right circumstances.

Bug: none
Test: gfxlatency

Change-Id: I30ec38d20a5f46e79418033f7ae63f095727740e