[cloud_provider_firestore] fix handling of object (over)write
It turns out we have an implicit contract with Ledger binary that
writing the same object twice must succeed (return OK). Without that we
fail the convergence benchmark and e2e test (yay for the tests catching
this!).

This patch aligns the firestore binary with this contract, adds docs and
a validation test; to unblock the switch.

In the next step we might want to re-evaluate this contract - might be
better to return ALREADY_EXISTS here and let the ledger binary pick the
policy of how to handle.

LE-540 #comment [cloud_provider_firestore] fix handling of object (over)write

Test: convergence sync benchmark, e2e sync tests w/ disabled conv test, validation test
Change-Id: I3790113997ae96af631b79012d076f1a23ef10c9