[ledger] Update FakeJournalDelegate to contain the value of the page
Before this CL FakeJournalDelegate contained the transformation done in
the commit.

This is problematic because fake_page_storage was not always correctly
handling this (in particular, GetEntryFromCommit was only looking to the
last commit).
This is also problematic once the Clear() operation is added on Commit,
as the transformation would then be dependant on the state of the
previous commit.

Instead, this CL replace FakeJournalDelegate so that it contains the
full content of the commit.

LE-537

TEST=Existing tests still pass.

Change-Id: Ibc8ad5404255d8ecef72ffca2d344768c414159a