[zircon][usb] Add protocol ops for usb-request functions
This changeset adds protocol ops for some of the functions in usb-request.c.
The motivation behind this changeset is to move the usb-request functionality out of
the ddk folder. Functionality pertaining to the usb operations is being moved to usb
protocol in this changeset and the rest will be moved to a system/ulib/driver/usb library.
I will modify the client calls in a later changeset.

Test: build. quick sanity test on vim2

Change-Id: Ia550e32b388c2757315f19b34f10b907f8a9d0d2