[dw][pcie] Pass ATU config as Device Metadata
The DesignWare PCI controller has an internal address
translation unit (iATU) that translates memory addresses
to PCI bus addresses.

For example, a memory read at address X might translate to
a PCI IO read for a particular device.

The memory addresses designated for the iATU differ by platform
and are only known by the board driver so we devise a scheme where
these ranges are passed as private device metadata to the
designware PCIe driver.

Test: Build and run on Gauss.
      Previously IO writes weren't generating IO TLPs on the PCIe bus
      which prevented drivers that relied on IO BARs from binding
      correctly. Verified that IO TLPs are now generated on the bus by
      reading data from drivers' IO space.

Change-Id: I8832415fe15578e0f7aca87e4a06decf5c79e421