[vulkan][swapchain] Don't use surface width and height
Having these is problematic because the surface is supposed
to live through resizing.

Some platforms return currentExtent by querying the underlying
window, but since imagePipe doesn't have a size, we just return
"size unknown" for currentExtent and expect that the application
creates the swapchain with imageExtent set to the scenic view
size.

Width and height will be removed from the vulkan surface struct
as part of the extension cleanup (MA-471).

MA-267 #comment

Test:
vim2: go/magma-tps#L1
Change-Id: I6b23423993e75ca1efca17fa42a8dae000f20bac