[zircon][usb] Use usb protocol ops instead of usb-request functions
In a previous changeset, I added protocol ops for the usb request functionality.
The motivation behind the changeset is to move usb-request.c out of ddk
and support the operations in usb protocol instead. This changeset, uses those
protocol ops instead of calling the functions in usb-request.c. In a
later changeset, I will move the usb-request.c out of the ddk folder.

Test: Build. Sanity test on vim2

ZX-2404 #comment Start using protocol ops

Change-Id: Ib3f0d3929138a09b23704fa0ee80ccf62a121d35