[garnet][usb drivers] Use usb protocol ops instead of usb-request functions
In a previous changeset, I added protocol ops for the usb request functionality.
The motivation behind the changeset is to move usb-request.c out of ddk
and support the operations in usb protocol instead. This changeset, uses those
protocol ops instead of calling the functions in usb-request.c. In a
later changeset, I will move the usb-request.c out of the ddk folder.

Test: fx full-build.

ZX-2404 #comment Start using protocol ops

Change-Id: Ife2cf48d78925e498608cbf42407ae637f3e60f8