[fidlmerge] add 'amendments' feature
This CL adds an option to fidlmerge that allows it to load a JSON
file of 'amendments' to the original FIDL JSON. For the time being,
the amendments are confined to 'exclusions', a list of declarations
that are to be excluded from the model.

This will be used for ostream formatter generation to exclude those
declarations for which formatters are hand-rolled.

TEST: Tested 'live' in build environment.

Change-Id: I9f4d2f6179675466b56748fe8c07172d66082a84