[fidlmerge, build] add new properties to fidlmerge_cpp.gni
This CL adds two new properties that can be used in appliciations of
the fidlmerge_cpp template. 'options' allows the existing --options
feature of fidlmerge to be used (basically a list of key/value
pairs). 'amendments_path' allows the new --amend feature of fidlmerge
to be used. This allows declarations to be excluded from the FIDL
model using a JSON file.

TEST: Tested manually...will soon be used in the build.

Change-Id: Ib714054b0134bc1d2ff2f0169a7a5c41742b1beb