[audio] Add a simple audio helper library.
Add a small library which implements most of the audio driver protocol
for simple audio drivers.  This is mostly common code which has been
cut and pasted 3 times so far (4 if you count the driver I was working
on when I did this, 5-6 if you count some of the drivers in the
pipeline)

I don't expect this code to last for a super long time.  When audio
drivers make the transition to FIDL, there may be a new library which
fills a similar role, but it will probably be implemented differently.
Alternatively, the majority of the logic implemented here may end up
getting pulled into the DDK/DevMgr.

Test: manual; integration with WiP VIM2 audio driver.
Change-Id: I04f6b81ddf08708a9211c2abb5b7c8c1d26f2985