[cros-ec][sensors] Relax ignoring of flags field
These flags provide side-information that we don't currently use
(tablet mode), but also is used to flag certain special packets like
timestamps and flushes.  These come with sensor_num 0, which may
correspond to a real sensor, so we need to ignore these.
The old code was overly aggressive and would cause us to not
receive sensor reports while in tablet mode.

Test: Confirmed reports now show up in tablet mode

SCN-856 #done

Change-Id: Ifaa48b82d839e59062960ce0762d9752d156b84c