[zircon][dev] Fallback RTC driver
This CL introduces a fake RTC driver which can be used on
any board which does not have an actual RTC chip.

The device does not save the last value set but even so
it is useful because the uper layers actually use
https://roughtime.googlesource.com/roughtime to set the
RTC at every boot, thus this device in that case is
as useful as a real RTC device.

TEST = this device is only bound on the Astro so boot
       up to garnet in the astro connected to the
       internet and via ssh issue $date and check is
       correct.

Change-Id: I95aafb9e425305067fbdb2ffba130537d2ec551d