[tools] Slightly modify publish-package.py for the SDK
Adds a new optional parameter for specifying the location of
host tools like pm, and plumbs through the variable that contains
the repo key directory so that a custom location can be
specified on the command line.

Test: Verified that main() still defaults to using build_dir for the
repo_key_dir when no --update-keys is specified. Also verified the
script takes pm and amber-publish from the host_x64 dir as before,
or uses the path passed in under --host-tools-dir.

Change-Id: Ib8c49c691316cf47d3849c67d1cc493bb61fc436