[escher] Add support for Escher host tests.
Escher-on-Linux uses the LunarG Vulkan SDK, which is
downloaded from Google Storage.  This is not easily
configured to run in CQ since environment variables
must be set, etc.  Therefore, the default is to use
a stub Vulkan implementation that allows everything
to be compiled; only those tests that don't require
Vulkan are run.  This is not ideal, but a major step
forward from previously, where Escher was not even
compiled on host, resulting in frustrating breakages.

TEST: escher_unittests
SCN-573 #done

Change-Id: Ia26f208cfd11e41e3c0a79e37a4e36b7aeae96aa