[debugger] Add code block searching.
Add variable printing in backtraces. There is currently no value
printing.

Add a "locals" command that finds and prints local variables (as for
parameters, these all currently say "TODO" as the value).

Fix a bug in DIE decoding where Decode would return false if nothing was
decoded. All the calling code treats this case as "corrupt" instead of
"nothing present", so this changes the decode to return true if nothing
was decoded but decoding worked.

TEST=unit

Change-Id: I45943fd13c0fd1ccfdbb454747a758907a09b68c