[debugger] Move common files to a separate dir.
The client directory previously contained two targets, the client
library (which depends on symbols), and the common library (which
both client and symbols depend on).

This patch moves that target to a separate directory under zxdb to make
this separation more clear. The client_object.* files were moved to the
client target which is where they should be (putting this in common was
a mistake).

No code change, only file moves and include path updates.

TEST=none

Change-Id: I49208717a300ab272420829938ef1881b5e22a9e