[mediaplayer] stop audiorenderer requesting packets too often
This CL fixes a bug in the audio renderer used the media player. A
'node' such as the audio renderer is supposed to request a packet
when it needs it and then stop asking until it needs another. Prior
to this CL, the audio renderer was requesting a packet repeatedly.

TEST: Ran stress test all night...no issues
Change-Id: I46be027ab70d27e027f3029022c39110015f48bc