[lib][debugger_utils] Use 'zx::job*' instead of non-const 'zx::job&' for args
The Google C++ style guide disallows using non-const reference types
for function, on the grounds that it makes the behaviour of a function
call site less obvious:
https://google.github.io/styleguide/cppguide.html#Reference_Arguments

The usual fix is to use a pointer type instead of a reference type.

For comparison with other cases, note that it is quite common to use a
non-const unique_ptr<T>* for returning results (or for passing a
object where ownership might optionally be transferred).

zx::job/zx::object is directly analogous to unique_ptr<T>; it just
holds a Zircon handle rather than a pointer.

Bug: none
Test: compile

Change-Id: I45f43dcabad177e25f9ce77d93c1bf0cea562dc4