[amlogic-video] Toward external frame allocation.
This changes h264_decoder to allocate output frames starting from VMOs
instead of starting with an io_buffer_t, and is part of allowing for an
external source of VMOs in which to store output frames.  The output
frame VMOs still have to be physically contiguous - this is checked for
by pre-existing code called by ConfigureCanvas(), and if they're not
contiguous, the error is plumbed to an error handler.

There is a very_temp_kludge_bti_handle in this CL which will go away
when BufferAllocator is wired up to call zx_vmo_create_contiguous()
instead (or possibly one of the participant drivers).

Some of this will help in plumbing the output video frame format
parameters.

Tested: use_h264_decoder_test and amlogic_video_decoder internal tests
Change-Id: I600a0e5db9f78fa61536a30e75c447eb890cd005