[suggestion_engine] Handle new types of interactions for interruptions
Also adds a filter for suggestions that are interrupting at the moment.
This way we mark interrupting suggestions and don't show them in next
spaces.

This was causing issues when an interruption appears on top of a place
where next suggestions appear. Both were showing up in the same space.
We only want to show the interruption until an action is done on it.

MI4-1211 #comment Handle new types of interactions for interruptions
MI4-1211 #done

TESTED=CQ, suggestion_engine_test, affected unittests, device

Change-Id: I99346662ccbc1398ef340ad8e3327450a4473618