[ethernet] Signal a status change on start
When a client connects to an ethernet device, set the status signal on
the rx fifo to encourage them to read the status. This is normally
triggered by the start() method in the ethernet protocol, except that
the new client isn't added to the active set at the time so it misses
the signal.

NET-713
Test: - manual testing in QEMU with the 'set_link' monitor command
      - /system/test/sys/ethernet-test

Change-Id: Ic330eda4de98dea9b7e0381e20e35c4532883051