[ethernet] Signal a status change on start
When a client connects to an ethernet device, set the status signal on
the rx fifo to encourage them to read the status. This is normally
triggered by the start() method in the ethernet protocol, except that
the new client isn't added to the active set at the time so it misses
the signal.

Reland of https://fuchsia-review.googlesource.com/c/zircon/+/181072

NET-713
Test: - manual testing in QEMU with the 'set_link' monitor command
      - /system/test/sys/ethernet-test
      - manual testing with a Pixelbook and a USB ethernet adapter

Change-Id: I36410293f3ca5225e90fea01468fa7a284548af1