[amber] introduce a json file based store
Combined with the change in the 3rd party repository, this drops the size of
the Amber binary from 7.2mb to 3.1mb.

The on-device tuf cache is lost during this transition, which is a deliberate
change. The source configuration file already contains the necessary root
keys to assert provenance, so no critical data is lost given the current
deployments - Amber will simply initialize a new local store.

Test: //garnet/go/src/amber:amber_source_test & manual

PKG-92 #done

Change-Id: Idfc01289bd3d7f2c6e39b9b32d2f5c7793de54dc