[gn][build] Pass `PATH` environment to build command of zircon
`create_gn_rules.py` runs `make` command with empty environment, and
the makefile runs `which` command to detect compilers.
However, the behavior of `which` differs on platforms. GNU which and
BSD which fails to find any command under the empty `PATH`, while
`which` in debianutils works well.

Currently, `create_gn_rules.py` depends on the Debian which behavior,
and has a workaround for Mac, that uses BSD which. But, it fails on
GNU which.

This CL whitelists `PATH` variable to pass through to makefile for
building zircon. So that, it works for GNU which.

Change-Id: Ief4c50227df3f9949cfc5d1754e95f786a8d5ac7