[amber] Deal with a corrupt 'update' package
The update package can be present on the device, but its contents
can be corrupt. This change treats this condition as the package
not existing instead of bailing out. This may result in a
re-apply of the latest system version, but is less bad then being
not able to update.

This issue has been particulary observed when people pave their
device with --no-data. The dynamic package index still has an
entry for the update package, but its blobs are gone.

PKG-165 #done

TEST=manual, test update was broken before and fixed after

Change-Id: I769e14853ce634bd47e744841ca44987fbe0bbf6