[cmx] Migrate some cmx files in cp to 'services'.
This migrates some low-hanging fruit to be service sandboxed, all tests.

This CL also features some minor code improvements, like printing the
component url when a component fails to connect and changing
NamespaceTest so that it won't trigger this warning simply because
it tried to connect to fuchsia.sys.Environment on startup.

TESTED=ran all garnet tests

Change-Id: I2716c07afc0754bcbad6ec577ac2596842748e19