[fidlgen] Make Go enums implement fmt.Stringer
Several format strings around the garnet source attempt to print out %s for
FIDL-generated enums. It seemed useful to be able to see the name of the
constant (instead of needing to cross-reference the source), so the enum
template is changed to have FIDL enums implement fmt.Stringer.
(Otherwise, the format strings don't match arguments; this fails to
compile with newer versions of Go.)

Test: manual

Change-Id: I6ce01d9a97cff402eff1831d39bf2ea7aa130d7d