[s912][audio] Correct the audio register range.
Correct the range for the audio out registers for the AmLogic S912.
The range starts at a 1k offset into a page and was asking for 4KB of
registers.  This caused the DDK to round down the start of the
requested range to a 4K boundary, then forced it to ask for 2 pages
worth of mapping (to cover the 4k starting from a 1k offset that we
had asked for).

Recently, something has become locked down and we no longer allow this
request (mostly, because it was wrong to begin with).  So...

Don't do this.  The entire audio out register block takes up a little
less then 1k, not 4k.  Ask for 1k instead, so we end up with just the
registers we want, and nothing extra.

Test: Manual testing on VIM2
Change-Id: I4452577ca3c47b01edfffae769d85d14bc055573