[uapp][usb-test-fwloader] Add updating usb-tester.
If there is no FX3 device connected, we try to look for a usb-tester
device, and set it to fwloader mode. This will cause it to re-enumerate
as a FX3 device and we can then load the new firmware.

TEST= plug in FX3 device
fx shell usb-test-fwloader firmware.img
fx shell usb-test-fwloader updated-firmware.img
runtests -t usb-test

Change-Id: If5ebc75461b14bd07ab911ea59fbc7155b3de336