[puppet_master] Bring story_id back to ExecuteResult
There are some places where we still depend on story_id:
- Story affinity in proposals is tied to story id.
- FocusProvider#FocusWatcher#FocusInfo provides story_id.

Those features are used in agents that receive puppet master. While we
work on a better solution, as part of the migration, it simplifies
things to be able to use both. Clients of puppet master will need to
keep this association.

TESTED=dispatch_story_command_executor_unittest

MI4-1185 #comment Bring story_id back to ExecuteResult

Change-Id: I4b35cda36bc92f55b2ea467a52e2420655a515ee