[debugger] Implement some expression evaluation.
Implements some initial parts of expression evaluation with tests. This
is not currently hooked up to the actual program, only test data
sources.

Identifiers and unary '-' are supported in the evaluator. Unary '-' is
not very important but is the simplest operation on a value so is a
good starting point for handling primitive types.

The primitive type support is currently very early and will require
significant updates as the capabilities are expanded.

TEST=unit

Change-Id: I67206fc749288e0086008c7918ee02db0ef20519