[media][audiocore] Rename audio_server
Ongoing changes that partition the audio subsystem into policy
and mechanism elements have already motivated us to rename the
'audio_server' interface to 'audio', which will be the outer,
policy-governing portion.  The inner interface (which will
only be called by the outward-facing policy layer surrounding
it) will be named 'audio_core'.

This CL takes the next step in the partitioning, renaming the
audio_server binary to audio_core.  As before, this binary
contains the inner, mechanism-oriented implementation of the
audio mixer and associated processing engine/pipeline.  (For
now, it also contains our outer-layer policy-oriented code,
but pulling that functionalitye into a separate 'audio' binary
can occur, as necessary, in the future.)

Test: existing tests (some of whom are also touched by this
renaming) all pass.

Change-Id: Ib4b6df680ab61b3f155640f85becb24da208b2fa