[kernel[rsrc] Fix size calculation when rounding MMIOs down
When a resource's mmio base is unaligned the validation check
page aligned it, but did not adjust the size to be larger
the same way it did if the size was unaligned. Now we adjust
the size based on the new base vs the old base.

The unaligned test has been updated to catch this because its
original parameters were not sufficient.

Test: core-tests

Change-Id: I364df5caef552dc19166f3696374ce5aec19d389