[zircon][usb-request]Move usb-request.c out of ulib/ddk
The motivation behind this changes is that eventually system/ulib/ddk
should go away and parts of it should move to where they belong.
This changeset moves usb-request.c, usb-request-test.c out of the ulib/ddk
to system/dev/lib/usb-request. This library is internal for system/dev/usb
and all its fucntionality will be exposed via the usb protocol and the
usb-function protocol to the client drivers.

Test: Build. Basic Sanity test on vim2. Ran usb-request-tests using
runtests /boot/test/ddk/driver-tests

Change-Id: Ia62693f51046b9c93e4a4b9abfd25286b0d9b972