[benchmarks] Use smaller mapping size for Mmu tests
This is in preparation for switching these test cases to use the
perftest library.

Before, Mmu/MapUnmapWithFaults is quite slow: it takes 40ms per test
run on an Intel NUC.  Since perftest currently does a fixed number of
runs of each test (1000 by default), that would make the test take a
long time (~40s).  (And with gbenchmark, it means we only get a small
number of runs of the test in the allotted time.)  Reducing the
mapping size makes the test a lot faster.

Bug: TO-651
Test: /pkgfs/packages/zircon_benchmarks/0/bin/app
Change-Id: I551dcc6371942135762f29e2030f934159704e0d