[benchmarks] Convert mmu.cc to use the perftest library
This means the test results will be logged to the Catapult dashboard.

Note that we lose the SetItemsProcessed() feature (which perftest does
not have), which calculated the average number of pages processed per
second.  However, we gain the ability to use multi-step tests, which
time the map, fault_in and unmap operations separately.

Bug: TO-651
Test: "/pkgfs/packages/zircon_benchmarks/0/bin/app -p"
Change-Id: I299c44b96f12f828ccaba37458180ca1c528d9b2