[musl] Fix alignment issues for tls on AArch64
musl had been assuming that alignment of TLS segments would
always be <= 16. This change removes use of TPOFF_K and
ensures the the offset of every tls segment is calculated
with the gap in mind. This change ensures that dso.tls.offset
is always the offset from the thread pointer itself rather
than attempting to be from the start of the tls block. This
is consistent (though backwards) with how things work on
x86-64.

Fixes ZX-1646

Test: boot/test/core/elf-tls-test
      runtests in garnet
      fixed nacl libc-tests in garnet

* elf-tls-test only tests the simplist access model
* it isn't clear runtests tests tls at all
* the fixed nacl libc-tests confirms that all initial-exec tests work

Change-Id: Ida3ff39a25adc7a1b5c85af7e643a44a4f748766