[cobalt] Change the Logger interface to adhere to the "Simple" layout.
- Replaced all string params with bounded-length strings
- Moved two of the methods out of the Logger interface into a second
interface called LoggerExt. This contains the methods for which there is
no natural way to express the parameters using the "Simple" layout.
- Created a third interface called LoggerLow. This contains versions of
the methods in LoggerExt that are expressed in the Simple layout, but
in a less natural way than in LoggerExt.

Additionally this CL contains some other miscellaneous things:

(a) In the interface LoggerLow I have the beginnings of some work towards
supporting a BulkLog() method. I did this only to record some thoughts
but this work is incomplete. I expect azani@ and prathyushk@ to take this
over.

(b) I added the ReleaseStage enum

(c) Fixed a copule of things that accidentally fell out of the previous CL.

TEST=compiles

Change-Id: I6146250f2e9e1a5e823b311634688a9bd29a513d