[overnet] Build out initial infrastructure
Completes messaging layer, adds a generic packet_link implementation.
Uses this packet_link for routable messages and for a UDP layer.
Adds packet_nub for multiplexing several links over a single endpoint,
and uses this for the UDP layer to multiplex over a single socket.

Introduce a public overnet API, and a tool for interacting with overnet
from the command line (both of these are essentially stubs right now).

Test: included
Change-Id: I7ee92a1a69b84f7c60ba4abc2025f09071cacbf3