[packages] Add a package_group template to refer to a set of packages
This adds a package_group template which can be referred to from a
product configuration just like a package() but expands out to a set of
packages based on a GN evaluation of the template. This enables
parameterizing the build based on conditional logic in GN, instead of
constructing JSON files for every permutation of packages we may care
about.

Test: CQ
BLD-240 #comment

Change-Id: I844b28cfe84ad11c0f3a4096c0c8ef0320fdd6d5